back_end/fastapi_server/requirements.txt
# brew update && brew install gcc (this includes gfortran)
scikit-learn>=1.0.1
pandas>=1.3.3
# python-dotenv >= 0.19.0
# gunicorn >= 20.1.0
uvicorn >= 0.14.0
fastapi >= 0.65.1
python-multipart
python-dateutil